Add task properties support
**Coordinator and Balancer breaking API changes**

Fixes #100

Tasks now can have extra metadata (properties) and internally are
repesented as an interface instead of a simple string ID.

The intention is to offer a minimal amount of metadata for balancers and
handlers to use when deciding what to do with a task. It is *not* meant
to be used as the canonical store for task configuration as there's no
support for changing the metadata and notifying handlers of changes.

We also don't want the coordinator to consider the broker a general
purpose database. Using it as a distributed lock is already difficult
enough without piling more complexity into it.

Also fixes a tiny obscure bug in fair balancer where it might try to
release the same task more than once per `Balance()` call.

**Client API changes**

* `Client.Nodes` was removed as it's not really useful enough to force
  clients to implement it.
* `Client.DeleteTask` was removed because it was dangerous and should
  *never* be used.
* `Client.SubmitTask` now accepts a properties map.